Add hook to only override service node creation
